Add compatibility table for OG DDLC
It is technically the same as the one for DDLC+, but this one was
constructed manually in-game. The numbers were later deduced from the
data set found in DDLC+, but only Sayori's numbers are complete because
there is no in-game way to know that words had several levels of
affinity.